modules moved to appropriate subpackages of soccer_learning
