+++ soccer_learning/agents/FoeQ.py
@@ -1,4 +1,4 @@
-from SoccerGameAgent import SoccerGameAgent
+from soccer_learning.agents import SoccerGameAgent
 import numpy as np
 from cvxopt import solvers, matrix
 import copy
