+++ soccer_learning/agents/__init__.py
@@ -1,0 +1,6 @@
+from soccer_learning.agents.CEQ import CEQ
+from soccer_learning.agents.FoeQ import FoeQ
+from soccer_learning.agents.FriendQ import FriendQ
+from soccer_learning.agents.QLearning import QLearning
+from soccer_learning.agents.RandomAgent import RandomAgent
+from soccer_learning.agents.SoccerGameAgent import SoccerGameAgent
