+++ soccer_learning/environments/__init__.py
@@ -1,0 +1,1 @@
+from soccer_learning.environments.SoccerEnv import SoccerEnv
