+++ soccer_learning/games/__init__.py
@@ -1,0 +1,1 @@
+from soccer_learning.games.SoccerGame import SoccerGame
